# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
servers:
  - url: http://localhost:8000

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorCode:
      type: string
      enum:
        - PRODUCT_NOT_FOUND
        - PRODUCT_INACTIVE
        - ORDER_NOT_FOUND
        - ORDER_LIMIT_EXCEEDED
        - ORDER_HAS_ACTIVE
        - INVALID_STATE_TRANSITION
        - INSUFFICIENT_STOCK
        - PROMO_CODE_INVALID
        - PROMO_CODE_MIN_AMOUNT
        - ORDER_OWNERSHIP_VIOLATION
        - VALIDATION_ERROR
        - ACCESS_DENIED
        - TOKEN_INVALID
        - TOKEN_EXPIRED
        - REFRESH_TOKEN_INVALID
        - USER_ALREADY_EXISTS
        - INVALID_CREDENTIALS

    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code: { $ref: "#/components/schemas/ErrorCode" }
        message: { type: string }
        details: { type: object, nullable: true, additionalProperties: true }

    Role: { type: string, enum: [USER, SELLER, ADMIN] }
    ProductStatus: { type: string, enum: [ACTIVE, INACTIVE, ARCHIVED] }
    OrderStatus: { type: string, enum: [CREATED, PAYMENT_PENDING, PAID, SHIPPED, COMPLETED, CANCELED] }
    DiscountType: { type: string, enum: [PERCENTAGE, FIXED_AMOUNT] }

    UserEntity:
      type: object
      required: [id, email, role, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email, maxLength: 320 }
        role: { $ref: "#/components/schemas/Role" }
        created_at: { type: string, format: date-time }

    ProductEntity:
      type: object
      required: [id, name, price, stock, category, status, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 255 }
        description: { type: string, nullable: true, maxLength: 4000 }
        price: { type: number, format: decimal, minimum: 0.01 }
        stock: { type: integer, minimum: 0 }
        category: { type: string, minLength: 1, maxLength: 100 }
        status: { $ref: "#/components/schemas/ProductStatus" }
        seller_id: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    PromoCodeEntity:
      type: object
      required: [id, code, discount_type, discount_value, min_order_amount, max_uses, current_uses, valid_from, valid_until, active]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        discount_type: { $ref: "#/components/schemas/DiscountType" }
        discount_value: { type: number, format: decimal, minimum: 0.01 }
        min_order_amount: { type: number, format: decimal, minimum: 0.01 }
        max_uses: { type: integer, minimum: 1 }
        current_uses: { type: integer, minimum: 0 }
        valid_from: { type: string, format: date-time }
        valid_until: { type: string, format: date-time }
        active: { type: boolean }

    OrderEntity:
      type: object
      required: [id, user_id, status, total_amount, discount_amount, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/OrderStatus" }
        promo_code_id: { type: string, format: uuid, nullable: true }
        total_amount: { type: number, format: decimal }
        discount_amount: { type: number, format: decimal }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    OrderItemEntity:
      type: object
      required: [id, order_id, product_id, quantity, price_at_order]
      properties:
        id: { type: string, format: uuid }
        order_id: { type: string, format: uuid }
        product_id: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1, maximum: 999 }
        price_at_order: { type: number, format: decimal }
        
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 8, maxLength: 128 }
        role: { $ref: "#/components/schemas/Role" }

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }

    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }

    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: bearer }
        expires_in: { type: integer, example: 1200 }

    ProductCreate:
      type: object
      required: [name, price, stock, category, status]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        description: { type: string, nullable: true, maxLength: 4000 }
        price: { type: number, format: decimal, minimum: 0.01 }
        stock: { type: integer, minimum: 0 }
        category: { type: string, minLength: 1, maxLength: 100 }
        status: { $ref: "#/components/schemas/ProductStatus" }

    ProductUpdate: { allOf: [ { $ref: "#/components/schemas/ProductCreate" } ] }

    ProductsPage:
      type: object
      required: [items, totalElements, page, size]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/ProductEntity" } }
        totalElements: { type: integer }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }

    OrderItemInput:
      type: object
      required: [product_id, quantity]
      properties:
        product_id: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1, maximum: 999 }

    CreateOrderRequest:
      type: object
      required: [items]
      properties:
        items: { type: array, minItems: 1, maxItems: 50, items: { $ref: "#/components/schemas/OrderItemInput" } }
        promo_code: { type: string, nullable: true }

    UpdateOrderRequest:
      type: object
      required: [items]
      properties:
        items: { type: array, minItems: 1, maxItems: 50, items: { $ref: "#/components/schemas/OrderItemInput" } }

    OrderResponse:
      type: object
      required: [order, items, promo_code]
      properties:
        order: { $ref: "#/components/schemas/OrderEntity" }
        items: { type: array, items: { $ref: "#/components/schemas/OrderItemEntity" } }
        promo_code:
          allOf:
            - $ref: "#/components/schemas/PromoCodeEntity"
          nullable: true

    PromoCodeCreate:
      type: object
      required: [code, discount_type, discount_value, min_order_amount, max_uses, valid_from, valid_until, active]
      properties:
        code: { type: string }
        discount_type: { $ref: "#/components/schemas/DiscountType" }
        discount_value: { type: number, format: decimal, minimum: 0.01 }
        min_order_amount: { type: number, format: decimal, minimum: 0.01 }
        max_uses: { type: integer, minimum: 1 }
        valid_from: { type: string, format: date-time }
        valid_until: { type: string, format: date-time }
        active: { type: boolean }

paths:
  /auth/register:
    post:
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/RegisterRequest" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/TokenPair" } } } }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "409": { description: Exists, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /auth/login:
    post:
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/LoginRequest" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/TokenPair" } } } }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Invalid, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /auth/refresh:
    post:
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/RefreshRequest" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/TokenPair" } } } }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }

  /products:
    get:
      security: [{ BearerAuth: [] }]
      description: "Allowed roles: USER, SELLER, ADMIN."
      x-roles-allowed: [USER, SELLER, ADMIN]
      parameters:
        - { in: query, name: page, schema: { type: integer, minimum: 0, default: 0 } }
        - { in: query, name: size, schema: { type: integer, minimum: 1, default: 20 } }
        - { in: query, name: status, schema: { $ref: "#/components/schemas/ProductStatus" } }
        - { in: query, name: category, schema: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ProductsPage" } } } }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    post:
      security: [{ BearerAuth: [] }]
      description: "Allowed roles: SELLER, ADMIN."
      x-roles-allowed: [SELLER, ADMIN]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ProductCreate" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/ProductEntity" } } } }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "403": { description: Forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }

  /products/{id}:
    get:
      security: [{ BearerAuth: [] }]
      description: "Allowed roles: USER, SELLER, ADMIN."
      x-roles-allowed: [USER, SELLER, ADMIN]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ProductEntity" } } } }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    put:
      security: [{ BearerAuth: [] }]
      description: "Allowed roles: SELLER, ADMIN. SELLER can modify only own products."
      x-roles-allowed: [SELLER, ADMIN]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ProductUpdate" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ProductEntity" } } } }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "403": { description: Forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    delete:
      security: [{ BearerAuth: [] }]
      description: "Allowed roles: SELLER, ADMIN. SELLER can archive only own products."
      x-roles-allowed: [SELLER, ADMIN]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        "204": { description: Archived }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "403": { description: Forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }

  /promo-codes:
    post:
      security: [{ BearerAuth: [] }]
      description: "Allowed roles: SELLER, ADMIN."
      x-roles-allowed: [SELLER, ADMIN]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PromoCodeCreate" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/PromoCodeEntity" } } } }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "403": { description: Forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }

  /orders:
    post:
      security: [{ BearerAuth: [] }]
      description: "Allowed roles: USER, SELLER, ADMIN."
      x-roles-allowed: [USER, SELLER, ADMIN]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateOrderRequest" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/OrderResponse" } } } }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Product not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "409": { description: Conflict, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "422": { description: Promo invalid, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "429": { description: Rate limit, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }

  /orders/{id}/cancel:
    post:
      security: [{ BearerAuth: [] }]
      description: "Allowed roles: USER, SELLER, ADMIN. Only owner can cancel order."
      x-roles-allowed: [USER, SELLER, ADMIN]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        "200": { description: Canceled, content: { application/json: { schema: { $ref: "#/components/schemas/OrderResponse" } } } }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "403": { description: Forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "409": { description: Conflict, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }

  /orders/{id}:
    put:
      security: [{ BearerAuth: [] }]
      description: "Allowed roles: USER, SELLER, ADMIN. Only owner can update order. Update allowed only in CREATED state."
      x-roles-allowed: [USER, SELLER, ADMIN]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/UpdateOrderRequest" } } } }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/OrderResponse" } } } }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "403": { description: Forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "409": { description: Conflict, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "429": { description: Rate limit, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }